BackEnd/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.39
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2
python-dotenv==1.0.0
mercadopago==2.0.6
requests==2.13.0
urllib3==1.26.20
python-jose[cryptography]==3.3.0
yagmail==0.15.293
python-multipart==0.0.6
fpdf==1.7.2
psycopg2-binary==2.8.6
supabase==2.0.0